feat: (commentary-form) move to standalone section, fix hx- action
