+++ .deco/blocks/pages-Blog%2520Post%2520Page-827555.json
@@ -23,6 +23,23 @@
       }
     },
     {
+      "__resolveType": "site/sections/BlogPostComment.tsx",
+      "page": {
+        "__resolveType": "blog/loaders/BlogPostPage.ts",
+        "slug": {
+          "__resolveType": "website/functions/requestToParam.ts",
+          "param": "slug"
+        }
+      },
+      "post": {
+        "__resolveType": "blog/loaders/BlogPostItem.ts",
+        "slug": {
+          "__resolveType": "website/functions/requestToParam.ts",
+          "param": "slug"
+        }
+      }
+    },
+    {
       "__resolveType": "site/sections/Newsletter.tsx"
     },
     {
