+++ sections/BlogPost.tsx
@@ -1,19 +1,11 @@
-import { AppContext } from "site/apps/site.ts";
-import type { AppContext as RecordsApp } from "site/apps/deco/records.ts";
 import { type BlogPost, BlogPostPage } from "apps/blog/types.ts";
-import { eq } from "drizzle-orm";
 import Image from "apps/website/components/Image.tsx";
-import { comments } from "site/db/schema.ts";
-import { CommentaryList } from "site/components/ui/CommentaryList.tsx";
-import type { UserCommentary } from "site/sdk/types.ts";
-import { CommentaryForm } from "site/components/ui/CommentaryForm.tsx";
 
 interface Props {
   /**
    * @description The description of name.
    */
   page?: BlogPostPage | null;
-  commentaries?: UserCommentary[];
 }
 
 const PARAGRAPH_STYLES = "[&_p]:leading-[150%] [&_*]:mb-4";
@@ -122,22 +114,7 @@ function SocialIcons() {
   );
 }
 
-export const loader = async (
-  props: Props,
-  _req: Request,
-  ctx: AppContext & RecordsApp,
-) => {
-  const drizzle = await ctx.invoke("records/loaders/drizzle.ts");
-
-  const recs = await drizzle
-    .select()
-    .from(comments)
-    .where(eq(comments.post_id, props.page!.post.name));
-
-  return { ...props, commentaries: recs };
-};
-
-export default function BlogPost({ page, commentaries }: Props) {
+export default function BlogPost({ page }: Props) {
   const { title, authors, image, date, content, name } = page?.post ||
     DEFAULT_PROPS;
 
@@ -220,8 +197,6 @@ export default function BlogPost({ page, commentaries }: Props) {
           </div>
         </div>
       </div>
-      <CommentaryList commentaries={commentaries} />
-      <CommentaryForm postId={name} />
     </div>
   );
 }
